Move `level` to top level
